chore(examples): websocket example
