--- examples/javascript/websocket/.dockerignore
@@ -1,13 +1,0 @@
-.git
-.vscode
-.idea
-node_modules
-.env
-npm-debug.log
-dist
-Dockerfile
-.dockerignore
-.DS_Store
-.env.*
-*.md
-!README.md
