+++ examples/javascript/websocket/package.json
@@ -1,25 +1,12 @@
 {
-	"name": "system-test-container",
+	"name": "example-system-test",
 	"version": "1.0.0",
 	"private": true,
-	"type": "module",
 	"dependencies": {
-		"@hono/node-server": "^1.13.8",
-		"@hono/node-ws": "^1.1.0",
+		"@rivet-gg/actor": "workspace:*",
 		"hono": "^4.6.17"
 	},
 	"devDependencies": {
-		"@rivet-gg/actor-core": "^5.1.2",
-		"@rivet-gg/api": "^24.6.2",
-		"@types/ws": "^8.18.0",
-		"node-fetch": "^3.3.2",
-		"tsx": "^4.7.0",
-		"typescript": "^5.3.3",
-		"ws": "^8.16.0"
-	},
-	"scripts": {
-		"start": "tsx ws.ts",
-		"test": "tsx tests/client.ts",
-		"build": "tsc --outDir dist"
+		"@rivet-gg/actor-client": "workspace:*"
 	}
 }
