+++ examples/javascript/websocket/rivet.json
@@ -1,7 +1,7 @@
 {
 	"builds": {
-		"main": {
-			"script": "src/main.ts",
+		"ws": {
+			"script": "ws.ts",
 			"access": "private"
 		}
 	},
