+++ examples/javascript/websocket/tsconfig.json
@@ -1,16 +1,13 @@
 {
 	"compilerOptions": {
-		"target": "ES2022",
-		"module": "NodeNext",
-		"moduleResolution": "NodeNext",
-		"types": ["node"],
+		"target": "ESNext",
+		"module": "ESNext",
+		"moduleResolution": "bundler",
+		"types": ["node", "deno"],
 		"allowSyntheticDefaultImports": true,
 		"esModuleInterop": true,
-		"isolatedModules": true,
-		"strict": true,
-		"skipLibCheck": true,
-		"outDir": "dist"
+		"noEmit": true,
+		"isolatedModules": true
 	},
-	"include": ["**/*.ts"],
-	"exclude": ["node_modules", "dist"]
+	"include": ["**/*.ts"]
 }
